components/embedding/requirements.txt
llama-cpp-python
pandas
numpy